Move device classes to models.py
