+++ homeassistant/components/amcrest/binary_sensor.py
@@ -41,7 +41,7 @@ from .const import (
 from .helpers import log_update_error, service_signal
 
 if TYPE_CHECKING:
-    from . import AmcrestConfiguredDevice, AmcrestDevice
+    from .models import AmcrestConfiguredDevice, AmcrestDevice
 
 
 @dataclass(frozen=True)
